+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - Added sql for monthly and weekly views
+- Added debug and account mapping scripts
 ### Changed
 ### Fixed
 - Fixed Duplicate issue in Materialized Views